config: extract flow/tcp-mss compilation to compiler_security_flow.go
Part B of #4121 (pure code motion). Moves the flow-compilation cluster
verbatim out of compiler_security.go, including the flow-traceoptions
commit-time gates and the tcp-mss range validators:

  - flowTraceFileNameError, the FlowTraceMin/Max size/count bounds
  - validateFlowTraceFileAST (#3420 path-traversal gate)
  - flowTraceImplementedFlags, validateFlowTraceFlagsAndFiltersAST
    (#3422 flag/filter gate)
  - flowTraceSizeFilesValues, validateFlowTraceSizeFilesAST
    (#3424 rotation-bounds gate)
  - tcpMSSKinds, validateTCPMSSRanges, checkTCPMSSKind
    (#1979/#2486 tcp-mss gates)
  - compileFlow (aging, tcp/udp/icmp session, tcp-mss, traceoptions)

Function bodies are byte-identical to origin/master. The new file
imports fmt, net/netip, path/filepath, strconv, strings (the union the
moved declarations use); net/netip, path/filepath, and strconv drop
out of the base import block. After this move the base
compiler_security.go holds only compileSecurity and imports fmt.

Validation: gofmt -l clean, go build ./pkg/config/ green.
